Create schema file for the api
