+++ person_image_segmentation/utils/schema.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+class PredictionResponse(BaseModel):
+    filename: str
+    message: str
+
+class ErrorResponse(BaseModel):
+    error: str
+
+class RootResponse(BaseModel):
+    message: str
